resetPassword

+++ src/Controller/UserController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\User;
 use App\Form\UserType;
+use App\Form\UserEditType;
 use App\Enum\Role;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -14,6 +15,7 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Authentication\UserAuthenticatorInterface;
 use Symfony\Component\Security\Core\Security;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 #[Route('/user')]
 class UserController extends AbstractController
@@ -21,21 +23,22 @@ class UserController extends AbstractController
     #[Route('/back', name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository, Security $security): Response
     {
-        $user=$this->getUser();
-        if ($user!==null) {
-           $role = $user->getRole();
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $user = $this->getUser();
+        if ($user !== null) {
+            $role = $user->getRole();
 
-           if ($role === Role::ADMIN) {
-            return $this->render('user/index.html.twig', [
-                'users' => $userRepository->findAll(),
-            ]);
-           } else {
-               // The user is a regular user
-               // Do something for regular users
-               return $this->redirectToRoute('app_user_index_front');
-           }
+            if ($role === Role::ADMIN) {
+                return $this->render('user/index.html.twig', [
+                    'users' => $userRepository->findAll(),
+                ]);
+            } else {
+                // The user is a regular user
+                // Do something for regular users
+                return $this->redirectToRoute('app_user_index_front');
+            }
         }
-        
+
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -43,8 +46,9 @@ class UserController extends AbstractController
     #[Route('/front', name: 'app_user_index_front', methods: ['GET'])]
     public function indexF(UserRepository $userRepository): Response
     {
-        $user=$this->getUser();
-         if ($user!==null) {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $user = $this->getUser();
+        if ($user !== null) {
             $role = $user->getRole();
 
             if ($role !== Role::ADMIN) {
@@ -56,15 +60,14 @@ class UserController extends AbstractController
                 // Do something for regular users
                 return $this->redirectToRoute('app_user_index');
             }
-         }
-        return $this->render('user/indexFront.html.twig', [
-            
-        ]);
+        }
+        return $this->redirectToRoute('app_login');
     }
 
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -85,6 +88,7 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -93,6 +97,7 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -111,11 +116,38 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
         }
 
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
+
+    #[Route('/{id}/editProfile', name: 'app_user_editProfile', methods: ['GET', 'POST'])]
+    public function editProfile(Request $request, User $utilisateur, EntityManagerInterface $entityManager, UserPasswordHasherInterface $userPasswordHasher): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $form = $this->createForm(UserEditType::class, $utilisateur);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $utilisateur->setPassword(
+                $userPasswordHasher->hashPassword(
+                    $utilisateur,
+                    $form->get('Password')->getData()
+                )
+            );
+            $entityManager->flush();
+
+            return $this->redirectToRoute('profile', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('user/editProfile.html.twig', [
+            'utilisateur' => $utilisateur,
+            'form' => $form,
+        ]);
+    }
+
 }
